+++ metadrive/component/waymo_block/waymo_block.py
@@ -9,7 +9,7 @@ from metadrive.constants import WaymoLaneProperty
 from metadrive.engine.engine_utils import get_engine
 from metadrive.utils.interpolating_line import InterpolatingLine
 from metadrive.utils.math_utils import wrap_to_pi, norm
-from metadrive.utils.waymo_utils.waymo_utils import convert_polyline_to_metadrive
+from metadrive.utils.waymo_utils.utils import convert_polyline_to_metadrive
 from metadrive.utils.waymo_utils import RoadLineType, RoadEdgeType
