+++ metadrive/obs/top_down_renderer.py
@@ -12,7 +12,7 @@ from metadrive.obs.top_down_obs_impl import WorldSurface, VehicleGraphics, LaneG
 from metadrive.utils.interpolating_line import InterpolatingLine
 from metadrive.utils.map_utils import is_map_related_instance
 from metadrive.utils.utils import import_pygame
-from metadrive.utils.waymo_utils.waymo_utils import convert_polyline_to_metadrive
+from metadrive.utils.waymo_utils.utils import convert_polyline_to_metadrive
 
 pygame = import_pygame()
