+++ metadrive/tests/test_functionality/test_read_waymo_data.py
@@ -1,5 +1,5 @@
 from metadrive.engine.asset_loader import AssetLoader
-from metadrive.utils.waymo_utils.waymo_utils import read_waymo_data
+from metadrive.utils.waymo_utils.utils import read_waymo_data
 
 
 def _test_read_waymo_data():
