+++ metadrive/utils/waymo_utils/__init__.py
@@ -1,2 +1,2 @@
-from metadrive.utils.waymo_utils.waymo_utils import read_waymo_data
+from metadrive.utils.waymo_utils.utils import read_waymo_data
 from metadrive.utils.waymo_utils.waymo_type import RoadLineType, RoadEdgeType
